Filter models based on both tag and protocol
Instead of filtering which models to learn based solely on the tag, it
is better to also include the protocol version in the comparison. The
reason is that some versions already contained one learned model, while
they supported more versions of TLS. These were not included in the
earlier check.