[CI SKIP] Manually update two models
There were two implementations (OpenSSL 1.0.1t and mbed TLS 1.3.3) for
which the model for TLS 1.2 was learned incorrectly. These
implementations had the exact same behavior as others
(indistinguishable), but contained more nodes and transitions.
Performing model learning again, resulted in the more minimal models).

It is not clear as to how this happened. It seems to be limited to these
two implementations, but this should be investigated more.